entity changed

+++ src/main/java/com/thecodereveal/shopease/repositories/RefundRequestRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
 
-public interface RefundRequestRepository extends JpaRepository<RefundRequest, Long> {
+public interface RefundRequestRepository extends JpaRepository<RefundRequest, Integer> {
 }
